% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_palettes.R
\docType{data}
\name{all_palettes}
\alias{all_palettes}
\title{All Palettes}
\format{
\subsection{\code{all_palettes}}{
}
}
\source{
\url{https://raw.githubusercontent.com/jwilber/Bob_Ross_Paintings/master/data/bob_ross_paintings.csv}
}
\usage{
all_palettes
}
\description{
This is the programmatically created palettes for each Bob Ross painting.
scale_fill_bob_ross, scale_color_bob_ross and show_colors_and_paintings all link to this data
Colors were mined leveraging a few image processing tools
imgpalr in particular was key for calculating the hex values
}
\keyword{datasets}
