% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bob_ross_paintings.R
\docType{data}
\name{bob_ross_paintings}
\alias{bob_ross_paintings}
\title{The Master Bob Ross Painting 'db'}
\format{
\subsection{\code{bob_ross_paintings}}{
}
}
\source{
\url{https://raw.githubusercontent.com/jwilber/Bob_Ross_Paintings/master/data/bob_ross_paintings.csv}
}
\usage{
bob_ross_paintings
}
\description{
sourced from a python library Bob_Ross_Paintings
}
\keyword{datasets}
