% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_bob_ross.R
\name{scale_fill_bob_ross}
\alias{scale_fill_bob_ross}
\title{Scale Fills, A ggplot2 helper function}
\usage{
scale_fill_bob_ross(
  painting = c("peaceful_valley", "a_walk_in_the_woods"),
  type = c("divergent", "qualitative")
)
}
\arguments{
\item{painting}{a vector of painting names}

\item{type}{divergent, qualitative, or both}
}
\value{
a set of hex values to scale colors on a ggplot2 object
}
\description{
Scale Fills, A ggplot2 helper function
}
\examples{
scale_fill_bob_ross(painting = "peaceful_valley", type = "qualitative")

}
