% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors_and_paintings.R
\name{show_colors_and_paintings}
\alias{show_colors_and_paintings}
\title{Show Colors & Paintings}
\usage{
show_colors_and_paintings(
  paintings = c("peaceful_valley", "a_walk_in_the_woods")
)
}
\arguments{
\item{paintings}{a vector of painting names from all_palettes}
}
\value{
a visualization of palettes and the painting
}
\description{
Show Colors & Paintings
}
\examples{
show_colors_and_paintings(paintings = c("peaceful_valley"))
}
