% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_bob_ross_colors.R
\docType{data}
\name{unique_bob_ross_colors}
\alias{unique_bob_ross_colors}
\title{Unique Bob Ross Colors}
\format{
\subsection{\code{unique_bob_ross_colors}}{
}
}
\source{
\url{https://raw.githubusercontent.com/jwilber/Bob_Ross_Paintings/master/data/bob_ross_paintings.csv}
}
\usage{
unique_bob_ross_colors
}
\description{
this lists all uniquely named colors and their hex values. These are the unblended unique colors
}
\keyword{datasets}
